Save message properties into @metadata fields.
Apart from the message payload itself, RabbitMQ messages contain
various metadata when they're delivered to a consumer. These can
include the timestamp the message was published, the name of the
user that published the message, the name of the exchange/queue
to which the message was published, etc.

Properties of the message and its delivery are stored in the
[@metadata][rabbitmq_properties] field while the user-defined
headers (technically a part of the properties) are stored in the
[@metadata][rabbitmq_headers] field.

Fixes #65