Docs: Rewrite overview paragraph and update option docs.
Since this plugin no longer uses the Bunny library the references
to it are deleted. The introduction also gets a description of the
newly added rabbitmq_properties and rabbitmq_headers @metadata
fields. Finally, several options had hard to understand or even
misleading descriptions, which have now been reworked and expanded.